db_final_backup/template/requirements.txt
typer>=0.9.0
rich>=13.5.0
sqlparse>=0.4.4
sqlglot>=19.8.0
matplotlib>=3.7.2
plotly>=5.16.0
pandas>=2.0.3
pydantic>=2.3.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.5  # PostgreSQL驱动
python-dotenv>=1.0.0
pyyaml>=6.0.1
transformers>=4.32.0
torch>=2.0.1  # 用于Hugging Face模型
dashscope>=1.12.0  # 阿里云Qwen API
deepseek>=0.1.0  # DeepSeek API库
networkx>=3.0  # 用于可视化查询计划
accelerate>=0.20.3  # 加速Hugging Face模型推理
llama-cpp-python>=0.2.0  # 可选，用于本地模型 